yaSSL (used in MySQL) returns unsupported cipher instead of Alert
yaSSL has a bug which will return a successful Server Hello message,
with a cipher that the server supports, even if none of the client
provided ciphers are supported.

This causes an infinite loop on many tests, which whittle down the
supported cipher list, by sending a full list in a loop, and
removing the successfully negotiated cipher from the list each time.

ex:
  run_pfs()
  cipher_pref_check()
  run_rc4()

This is a proposed fix within run_rc4() that should apply to the other
situations as well.