LibreSSL doesn't support "engine"
LibreSSL prints an "invalid command" error for "openssl engine ..." regardless of what "..." is, but always returns 0. The result is that when using LibreSSL, the initialize_engine() concludes that there is engine support even though there isn't. This results in error messages later on in testing.

This PR fixes the problem by checking whether "$OPENSSL engine -v" results in a "invalid command" error being send to stderr before trying any other tests with engine.

On the system in which I was testing, fixing the above problem still lead to a few places in which $OPENSSL printed a "can't open config file" warning, where the warning was not being redirected to /dev/null or to $ERRFILE. This PR also fixes that.