Fix socksend() debugging output.
For debug levels 4 and above, socksend() is supposed to print the data that is being sent to the terminal. It currently uses "echo -e" for this so that a newline character can be written before the data. However, the "-e" option causes every byte of $data, written as "\xHH" to be interpreted as an escape sequence rather than just being printed as plain text.

This PR fixes the problem by removing the "-e" option and using a separate call to "echo" to create the preceding newline character.